change: move follow list to navigator settings
