+++ src/components/NavigatorSettings.tsx
@@ -1,9 +1,30 @@
-import { useState } from 'react'
-import { Divider, Tabs, Box, Tab, Typography } from '@mui/material'
+import { useContext, useEffect, useState } from 'react'
+import {
+    Divider,
+    Tabs,
+    Box,
+    Tab,
+    Typography,
+    List,
+    ListItem,
+    Button,
+    ListItemButton,
+    ListItemAvatar,
+    Avatar,
+    ListItemText
+} from '@mui/material'
 import { StreamPicker } from './StreamPicker'
 import { usePersistent } from '../hooks/usePersistent'
+import type { User } from '../model'
+import { ApplicationContext } from '../App'
 
-export function NavigatorSettings(): JSX.Element {
+export interface NavigatorSettingsProps {
+    followList: string[]
+    setFollowList: (newlist: string[]) => void
+}
+
+export function NavigatorSettings(props: NavigatorSettingsProps): JSX.Element {
+    const appData = useContext(ApplicationContext)
     const [tab, setTab] = useState(0)
 
     const [followStreams, setFollowStreams] = usePersistent<string[]>(
@@ -18,6 +39,20 @@ export function NavigatorSettings(): JSX.Element {
         'defaultPostNonHome',
         []
     )
+
+    const [followList, setFollowList] = useState<User[]>([])
+
+    useEffect(() => {
+        Promise.all(
+            props.followList.map((ccaddress) => appData.userDict.get(ccaddress))
+        ).then((e) => {
+            setFollowList(e)
+        })
+    }, [props.followList])
+
+    const unfollow = (ccaddress: string): void => {
+        props.setFollowList(props.followList.filter((e) => e !== ccaddress))
+    }
 
     return (
         <Box>
@@ -35,6 +70,38 @@ export function NavigatorSettings(): JSX.Element {
             {tab === 0 && (
                 <Box>
                     <Typography variant="h3">ユーザー</Typography>
+                    <List
+                        dense
+                        sx={{
+                            width: '100%',
+                            maxWidth: 360,
+                            maxHeight: '300px',
+                            bgcolor: 'background.paper'
+                        }}
+                    >
+                        {followList.map((user) => (
+                            <ListItem
+                                key={user.ccaddress}
+                                secondaryAction={
+                                    <Button
+                                        onClick={() => {
+                                            unfollow(user.ccaddress)
+                                        }}
+                                    >
+                                        unfollow
+                                    </Button>
+                                }
+                                disablePadding
+                            >
+                                <ListItemButton>
+                                    <ListItemAvatar>
+                                        <Avatar src={user.avatar} />
+                                    </ListItemAvatar>
+                                    <ListItemText primary={user.username} />
+                                </ListItemButton>
+                            </ListItem>
+                        ))}
+                    </List>
                     <Divider />
                     <Typography variant="h3">ストリーム</Typography>
                     <StreamPicker
